# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/account/balance": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "查询用户账户的余额",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "account"
                ],
                "summary": "查询用户余额",
                "parameters": [
                    {
                        "type": "string",
                        "description": "JWT token",
                        "name": "Access-Token",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "$ref": "#/definitions/resp.Response"
                        }
                    },
                    "500": {
                        "description": "内部错误",
                        "schema": {
                            "$ref": "#/definitions/resp.ErrResponse"
                        }
                    }
                }
            }
        },
        "/account/withdraw": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "处理用户提现请求,减少用户余额,发起转账",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "account"
                ],
                "summary": "用户提现",
                "parameters": [
                    {
                        "description": "提现信息",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/account.Amount"
                        }
                    },
                    {
                        "type": "string",
                        "description": "JWT token",
                        "name": "Access-Token",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "$ref": "#/definitions/resp.Response"
                        }
                    },
                    "400": {
                        "description": "请求错误",
                        "schema": {
                            "$ref": "#/definitions/resp.ErrResponse"
                        }
                    },
                    "500": {
                        "description": "内部错误",
                        "schema": {
                            "$ref": "#/definitions/resp.ErrResponse"
                        }
                    }
                }
            }
        },
        "/account/withdrawal-record": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "获取用户的提现记录",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "account"
                ],
                "summary": "用户提现记录",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "页码",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "每页数量",
                        "name": "page_size",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "JWT token",
                        "name": "Access-Token",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "$ref": "#/definitions/utils.PaginationQ"
                        }
                    },
                    "500": {
                        "description": "内部错误",
                        "schema": {
                            "$ref": "#/definitions/resp.ErrResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "account.Amount": {
            "type": "object",
            "required": [
                "amount"
            ],
            "properties": {
                "amount": {
                    "type": "number"
                }
            }
        },
...
        "resp.ErrResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "default": 1000
                },
                "msg": {
                    "type": "string"
                },
                "time_now": {
                    "type": "integer"
                }
            }
        },
        "resp.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "default": 0
                },
                "data": {},
                "time_now": {
                    "type": "integer"
                }
            }
        },
...
        "utils.PaginationQ": {
            "type": "object",
            "properties": {
                "list": {
                    "description": "save pagination list"
                },
                "page": {
                    "type": "integer"
                },
                "size": {
                    "description": "Ok   bool   `json:\"ok\"`",
                    "type": "integer"
                }
            }
        }
    }
}